moves locales back to scripts.js (#170)
* moves locales back to scripts.js

* delete locales.js

* inline LOCALES into paths-config.js, remove scripts.js export
