+++ tools/generator/paths-config.js
@@ -1,4 +1,98 @@
-import LOCALES from '../../scripts/locales.js';
+const LOCALES = {
+  '': { ietf: 'en-US', tk: 'hah7vzn.css' },
+  ae_ar: { ietf: 'ar', tk: 'qxw8hzm.css', dir: 'rtl' },
+  ae_en: { ietf: 'en', tk: 'hah7vzn.css', base: '' },
+  africa: { ietf: 'en', tk: 'hah7vzn.css', base: '' },
+  ar: { ietf: 'es-AR', tk: 'hah7vzn.css', exl: 'es' },
+  at: { ietf: 'de-AT', tk: 'hah7vzn.css', exl: 'de', base: 'de' },
+  au: { ietf: 'en-AU', tk: 'hah7vzn.css' },
+  be_en: { ietf: 'en-BE', tk: 'hah7vzn.css', base: '' },
+  be_fr: { ietf: 'fr-BE', tk: 'hah7vzn.css', exl: 'fr', base: 'fr' },
+  be_nl: { ietf: 'nl-BE', tk: 'qxw8hzm.css', exl: 'nl' },
+  bg: { ietf: 'bg-BG', tk: 'qxw8hzm.css', base: '' },
+  br: { ietf: 'pt-BR', tk: 'hah7vzn.css', exl: 'pt-br', base: 'pt' },
+  ca_fr: { ietf: 'fr-CA', tk: 'hah7vzn.css', exl: 'fr', base: 'fr' },
+  ca: { ietf: 'en-CA', tk: 'hah7vzn.css', base: '' },
+  ch_de: { ietf: 'de-CH', tk: 'hah7vzn.css', exl: 'de', base: 'de' },
+  ch_fr: { ietf: 'fr-CH', tk: 'hah7vzn.css', exl: 'fr', base: 'fr' },
+  ch_it: { ietf: 'it-CH', tk: 'hah7vzn.css', exl: 'it', base: 'it' },
+  cl: { ietf: 'es-CL', tk: 'hah7vzn.css', exl: 'es' },
+  cn: { ietf: 'zh-CN', tk: 'qxw8hzm', exl: 'zh-hans', base: '' },
+  co: { ietf: 'es-CO', tk: 'hah7vzn.css', exl: 'es' },
+  cr: { ietf: 'es-419', tk: 'hah7vzn.css' },
+  cy_en: { ietf: 'en-CY', tk: 'hah7vzn.css' },
+  cz: { ietf: 'cs-CZ', tk: 'qxw8hzm.css', base: '' },
+  de: { ietf: 'de-DE', tk: 'hah7vzn.css', exl: 'de' },
+  dk: { ietf: 'da-DK', tk: 'qxw8hzm.css', base: '' },
+  ec: { ietf: 'es-419', tk: 'hah7vzn.css' },
+  ee: { ietf: 'et-EE', tk: 'qxw8hzm.css', base: '' },
+  eg_ar: { ietf: 'ar', tk: 'qxw8hzm.css', dir: 'rtl' },
+  eg_en: { ietf: 'en-GB', tk: 'hah7vzn.css' },
+  el: { ietf: 'el', tk: 'qxw8hzm.css' },
+  es: { ietf: 'es-ES', tk: 'hah7vzn.css', exl: 'es' },
+  fi: { ietf: 'fi-FI', tk: 'qxw8hzm.css', base: '' },
+  fr: { ietf: 'fr-FR', tk: 'hah7vzn.css', exl: 'fr' },
+  gr_el: { ietf: 'el', tk: 'qxw8hzm.css' },
+  gr_en: { ietf: 'en-GR', tk: 'hah7vzn.css', base: '' },
+  gt: { ietf: 'es-419', tk: 'hah7vzn.css' },
+  hk_en: { ietf: 'en-HK', tk: 'hah7vzn.css', base: '' },
+  hk_zh: { ietf: 'zh-HK', tk: 'jay0ecd', exl: 'zh-hant' },
+  hu: { ietf: 'hu-HU', tk: 'qxw8hzm.css' },
+  id_en: { ietf: 'en', tk: 'hah7vzn.css', base: '' },
+  id_id: { ietf: 'id', tk: 'qxw8hzm.css' },
+  ie: { ietf: 'en-GB', tk: 'hah7vzn.css', base: '' },
+  il_en: { ietf: 'en-IL', tk: 'hah7vzn.css', base: '' },
+  il_he: { ietf: 'he', tk: 'qxw8hzm.css', dir: 'rtl' },
+  in_hi: { ietf: 'hi', tk: 'qxw8hzm.css' },
+  in: { ietf: 'en-IN', tk: 'hah7vzn.css' },
+  it: { ietf: 'it-IT', tk: 'hah7vzn.css', exl: 'it' },
+  jp: { ietf: 'ja-JP', tk: 'dvg6awq', exl: 'ja' },
+  kr: { ietf: 'ko-KR', tk: 'qjs5sfm', exl: 'ko' },
+  kw_ar: { ietf: 'ar', tk: 'qxw8hzm.css', dir: 'rtl' },
+  kw_en: { ietf: 'en-GB', tk: 'hah7vzn.css' },
+  la: { ietf: 'es-LA', tk: 'hah7vzn.css', exl: 'es', base: 'es' },
+  langstore: { ietf: 'en-US', tk: 'hah7vzn.css' },
+  lt: { ietf: 'lt-LT', tk: 'qxw8hzm.css' },
+  lu_de: { ietf: 'de-LU', tk: 'hah7vzn.css', exl: 'de', base: 'de' },
+  lu_en: { ietf: 'en-LU', tk: 'hah7vzn.css', base: '' },
+  lu_fr: { ietf: 'fr-LU', tk: 'hah7vzn.css', exl: 'fr', base: 'fr' },
+  lv: { ietf: 'lv-LV', tk: 'qxw8hzm.css' },
+  mena_ar: { ietf: 'ar', tk: 'qxw8hzm.css', dir: 'rtl' },
+  mena_en: { ietf: 'en', tk: 'hah7vzn.css', base: '' },
+  mt: { ietf: 'en-MT', tk: 'hah7vzn.css' },
+  mx: { ietf: 'es-MX', tk: 'hah7vzn.css', exl: 'es' },
+  my_en: { ietf: 'en-GB', tk: 'hah7vzn.css', base: '' },
+  my_ms: { ietf: 'ms', tk: 'qxw8hzm.css' },
+  ng: { ietf: 'en-GB', tk: 'hah7vzn.css' },
+  nl: { ietf: 'nl-NL', tk: 'qxw8hzm.css', exl: 'nl', base: '' },
+  no: { ietf: 'no-NO', tk: 'qxw8hzm.css', base: '' },
+  nz: { ietf: 'en-GB', tk: 'hah7vzn.css', base: '' },
+  pe: { ietf: 'es-PE', tk: 'hah7vzn.css', exl: 'es' },
+  ph_en: { ietf: 'en', tk: 'hah7vzn.css', base: '' },
+  ph_fil: { ietf: 'fil-PH', tk: 'qxw8hzm.css' },
+  pl: { ietf: 'pl-PL', tk: 'qxw8hzm.css', base: '' },
+  pr: { ietf: 'es-419', tk: 'hah7vzn.css' },
+  pt: { ietf: 'pt-PT', tk: 'hah7vzn.css', exl: 'pt-br' },
+  qa_ar: { ietf: 'ar', tk: 'qxw8hzm.css', dir: 'rtl' },
+  qa_en: { ietf: 'en-GB', tk: 'hah7vzn.css' },
+  ro: { ietf: 'ro-RO', tk: 'qxw8hzm.css', base: '' },
+  ru: { ietf: 'ru-RU', tk: 'qxw8hzm.css', base: '' },
+  sa_ar: { ietf: 'ar', tk: 'qxw8hzm.css', dir: 'rtl' },
+  sa_en: { ietf: 'en', tk: 'hah7vzn.css', base: '' },
+  se: { ietf: 'sv-SE', tk: 'qxw8hzm.css', exl: 'sv', base: '' },
+  sg: { ietf: 'en-SG', tk: 'hah7vzn.css', base: '' },
+  si: { ietf: 'sl-SI', tk: 'qxw8hzm.css', base: '' },
+  sk: { ietf: 'sk-SK', tk: 'qxw8hzm.css', base: '' },
+  th_en: { ietf: 'en', tk: 'hah7vzn.css', base: '' },
+  th_th: { ietf: 'th', tk: 'lqo2bst.css' },
+  tr: { ietf: 'tr-TR', tk: 'qxw8hzm.css', base: '' },
+  tw: { ietf: 'zh-TW', tk: 'jay0ecd', exl: 'zh-hant', base: '' },
+  ua: { ietf: 'uk-UA', tk: 'qxw8hzm.css', base: '' },
+  uk: { ietf: 'en-GB', tk: 'hah7vzn.css' },
+  vn_en: { ietf: 'en-GB', tk: 'hah7vzn.css', base: '' },
+  vn_vi: { ietf: 'vi', tk: 'qxw8hzm.css' },
+  za: { ietf: 'en-GB', tk: 'hah7vzn.css' },
+};
 
 export const ORG = 'adobecom';
 export const REPO = 'da-bacom';
